fix convoluted tabs
