+++ src/lib/components/response-pane/response-pane.svelte
@@ -5,13 +5,13 @@
     import { json } from "@codemirror/lang-json";
 
     import { Tabs, TabsList, TabsTrigger, TabsContent } from "$lib/components/primitives/tabs";
-    import { Alert } from "$lib/components/primitives/alert";
     import { CodeBlock } from "$lib/components/code-block";
     import { Button } from "$lib/components/primitives/button";
     import { Badge } from "$lib/components/primitives/badge";
     import { HTTP_STATUS_DESCRIPTION } from "$lib/utils/constants";
     import type { ActiveRequest } from "$lib/models";
     import type { HttpRes, SpaceCookie } from "$lib/bindings";
+    import { prettyJson, formatSize, formatElapsed } from "$lib/utils";
 
     type Props = {
         pane: PaneAPI;
@@ -20,56 +20,6 @@
     };
 
     let { pane, isCollapsed, activeReq }: Props = $props();
-
-    function prettyJson(data: string | undefined) {
-        if (!data) return String();
-
-        try {
-            return JSON.stringify(JSON.parse(data), null, 2);
-        } catch {
-            return data;
-        }
-    }
-
-    function formatElapsed(ms: number): string {
-        if (ms < 1000) return `${ms} ms`;
-
-        const seconds = ms / 1000;
-        if (seconds < 60) {
-            return seconds % 1 === 0
-                ? `${seconds}s`
-                : `${seconds.toFixed(2).replace(/\.?0+$/, "")} s`;
-        }
-
-        const minutes = Math.floor(seconds / 60);
-        const secRemainder = Math.floor(seconds % 60);
-        if (minutes < 60) {
-            return `${minutes} m ${secRemainder} s`;
-        }
-
-        const hours = Math.floor(minutes / 60);
-        const minRemainder = minutes % 60;
-
-        return `${hours} h ${minRemainder} m`;
-    }
-
-    function formatSize(bytes: number): string {
-        if (bytes < 1024) return `${bytes} B`;
-
-        const kb = bytes / 1024;
-        if (kb < 1024) {
-            return kb % 1 === 0 ? `${kb} KB` : `${kb.toFixed(2).replace(/\.?0+$/, "")} KB`;
-        }
-
-        const mb = kb / 1024;
-        if (mb < 1024) {
-            return mb % 1 === 0 ? `${mb} MB` : `${mb.toFixed(2).replace(/\.?0+$/, "")} MB`;
-        }
-
-        const gb = mb / 1024;
-
-        return gb % 1 === 0 ? `${gb} GB` : `${gb.toFixed(2).replace(/\.?0+$/, "")} GB`;
-    }
 </script>
 
 {#snippet httpResMeta(httpRes: HttpRes)}
@@ -158,8 +108,19 @@
     </div>
 {/snippet}
 
-{#snippet responseBtn(collapsed: boolean, onclick: () => void)}
-    <Button variant="ghost" {onclick} class="mr-1 hover:bg-transparent">
+{#snippet responseBtn(collapsed: boolean)}
+    <Button
+        variant="ghost"
+        onclick={() => {
+            if (isCollapsed) {
+                pane.expand();
+                pane.resize(60);
+            } else {
+                pane.collapse();
+            }
+        }}
+        class="mr-1 hover:bg-transparent"
+    >
         <span class="pr-1.5 text-xs font-medium">Response</span>
         {#if collapsed}
             <ChevronUpIcon size={14} />
@@ -169,178 +130,132 @@
     </Button>
 {/snippet}
 
-<div class="size-full">
-    {#if activeReq.self.status === "Idle"}
+{#snippet responseBar()}
+    <div class="bg-card flex h-8 w-full items-center justify-between border-y border-t-transparent">
         {#if isCollapsed}
-            <div class="bg-accent/25 flex h-8 w-full items-center justify-end border-b">
-                {@render responseBtn(isCollapsed, () => {
-                    pane.expand();
-                    pane.resize(60);
-                })}
+            <div class="flex h-8 w-full items-center justify-end gap-1.5 border-b">
+                {#if (activeReq.self.status === "Success" || activeReq.self.status === "Error") && activeReq.self.response}
+                    {@render httpResMeta(activeReq.self.response)}
+                {/if}
+                {@render responseBtn(isCollapsed)}
             </div>
         {:else}
-            <div class="bg-accent/25 flex h-8 w-full items-center justify-end border-b">
-                {@render responseBtn(isCollapsed, () => {
-                    pane.collapse();
-                })}
-            </div>
-            <div class="flex size-full items-center justify-center gap-2 pb-8">
-                <RocketIcon size="20" />
-                <span>
-                    Hit <b class="font-semibold">Send</b> to make a request
-                </span>
-            </div>
-        {/if}
-    {:else if activeReq.self.status === "Pending"}
-        <div class="flex size-full items-center justify-center">
-            <RefreshCwIcon
-                strokeWidth={1.5}
-                absoluteStrokeWidth
-                size={20}
-                class="mr-3 animate-spin"
-            />
-        </div>
-    {:else}
-        <Tabs value="body" class="size-full">
-            <div
-                class="bg-card flex h-8 w-full items-center justify-between border-y border-t-transparent"
-            >
-                {#if isCollapsed}
-                    <div class="flex h-8 w-full items-center justify-end gap-1.5 border-b">
-                        {#if activeReq.self.response}
-                            {@render httpResMeta(activeReq.self.response)}
-                        {/if}
-                        {@render responseBtn(isCollapsed, () => {
-                            pane.expand();
-                            pane.resize(60);
-                        })}
-                    </div>
-                {:else}
-                    <div class="px-1.5">
-                        <TabsList
-                            class="grid auto-cols-min grid-flow-col justify-start gap-2 p-0 [&>*]:text-xs"
-                        >
-                            <TabsTrigger value="body">Body</TabsTrigger>
-                            <TabsTrigger value="cookies">
-                                {"Cookies".concat(
-                                    activeReq.self.response?.cookies
-                                        ? ` (${activeReq.self.response?.cookies.length})`
-                                        : "",
-                                )}
-                            </TabsTrigger>
-                            <TabsTrigger value="headers">
-                                {"Headers".concat(
-                                    activeReq.self.response?.headers
-                                        ? ` (${activeReq.self.response?.headers.length})`
-                                        : "",
-                                )}
-                            </TabsTrigger>
-                        </TabsList>
-                    </div>
-                    <div class="flex h-8 w-full items-center justify-end gap-1.5 border-b">
-                        {#if activeReq.self.response}
-                            {@render httpResMeta(activeReq.self.response)}
-                        {/if}
-                        {@render responseBtn(isCollapsed, () => {
-                            pane.collapse();
-                        })}
-                    </div>
-                {/if}
-            </div>
-            {#if !isCollapsed}
-                <div class="bg-background flex h-[calc(100%-32px)] w-full">
-                    <TabsContent value="body" class="m-0 size-full">
-                        {#if activeReq.self.status === "Success"}
-                            <Tabs value="pretty" class="bg-card size-full">
-                                <div class="flex items-center justify-end border-b px-3">
-                                    <TabsList class="my-1 auto-cols-min grid-flow-col gap-2 p-0">
-                                        <TabsTrigger value="pretty">Pretty</TabsTrigger>
-                                        <TabsTrigger value="raw">Raw</TabsTrigger>
-                                        <TabsTrigger value="preview">Preview</TabsTrigger>
-                                    </TabsList>
-                                </div>
-                                <div
-                                    class="h-[calc(100%-2.25rem)] w-full overflow-scroll [&>*]:m-0"
-                                >
-                                    <TabsContent value="pretty" class="size-full">
-                                        <CodeBlock
-                                            language={json()}
-                                            readOnly={true}
-                                            value={prettyJson(activeReq.self.response?.data)}
-                                            class="size-full"
-                                        />
-                                    </TabsContent>
-                                    <TabsContent value="raw" class="size-full">
-                                        <CodeBlock
-                                            language={null}
-                                            readOnly={true}
-                                            value={activeReq.self.response?.data}
-                                            class="size-full"
-                                        />
-                                    </TabsContent>
-                                    <TabsContent value="preview" class="size-full">
-                                        <iframe
-                                            title=""
-                                            src="about:blank"
-                                            srcdoc={activeReq.self.response
-                                                ? activeReq.self.response.data
-                                                : ""}
-                                            class="size-full"
-                                            loading="lazy"
-                                            sandbox=""
-                                        ></iframe>
-                                    </TabsContent>
-                                </div>
-                            </Tabs>
-                        {:else if activeReq.self.status === "Error"}
-                            {#if activeReq.self.response && activeReq.self.response.data}
-                                <div class="flex size-full items-center justify-center gap-2">
-                                    <Alert
-                                        variant="destructive"
-                                        class="w-fit max-w-[50%] py-1 [&>*]:select-text"
-                                    >
-                                        <span>{activeReq.self.response.data}</span>
-                                    </Alert>
-                                </div>
-                            {:else}
-                                <div
-                                    class="h-[calc(100%-2.25rem)] w-full overflow-scroll [&>*]:m-0"
-                                >
-                                    <CodeBlock
-                                        language={json()}
-                                        readOnly={true}
-                                        value={activeReq.self.response &&
-                                        activeReq.self.response.status
-                                            ? HTTP_STATUS_DESCRIPTION[
-                                                  activeReq.self.response.status
-                                              ]
-                                            : "Something went wrong."}
-                                        class="size-full"
-                                    />
-                                </div>
-                            {/if}
-                        {/if}
-                    </TabsContent>
-                    <TabsContent value="cookies" class="m-0 size-full">
-                        {#if activeReq.self.response}
-                            {@render cookiesTable(activeReq.self.response.cookies)}
-                        {:else}
-                            <div class="flex size-full items-center justify-center">
-                                No cookies for you :(
-                            </div>
-                        {/if}
-                    </TabsContent>
-                    <TabsContent value="headers" class="m-0 size-full">
-                        {#if activeReq.self.response}
-                            {@render headersTable(activeReq.self.response.headers)}
-                        {:else}
-                            <div class="flex size-full items-center justify-center">
-                                No headers received
-                            </div>
-                        {/if}
-                    </TabsContent>
+            {#if activeReq.self.status === "Success" || activeReq.self.status === "Error"}
+                <div class="px-1.5">
+                    <TabsList
+                        class="grid auto-cols-min grid-flow-col justify-start gap-2 p-0 [&>*]:text-xs"
+                    >
+                        <TabsTrigger value="body">Body</TabsTrigger>
+                        <TabsTrigger value="cookies">
+                            {"Cookies".concat(
+                                activeReq.self.response?.cookies
+                                    ? ` (${activeReq.self.response?.cookies.length})`
+                                    : "",
+                            )}
+                        </TabsTrigger>
+                        <TabsTrigger value="headers">
+                            {"Headers".concat(
+                                activeReq.self.response?.headers
+                                    ? ` (${activeReq.self.response?.headers.length})`
+                                    : "",
+                            )}
+                        </TabsTrigger>
+                    </TabsList>
                 </div>
             {/if}
-        </Tabs>
+
+            <div class="flex h-8 w-full items-center justify-end gap-1.5 border-b">
+                {#if (activeReq.self.status === "Success" || activeReq.self.status === "Error") && activeReq.self.response}
+                    {@render httpResMeta(activeReq.self.response)}
+                {/if}
+                {@render responseBtn(isCollapsed)}
+            </div>
+        {/if}
+    </div>
+{/snippet}
+
+<Tabs value="body" class="size-full">
+    {@render responseBar()}
+
+    {#if !isCollapsed}
+        <div class="bg-background flex h-[calc(100%-32px)] w-full">
+            <TabsContent value="body" class="m-0 size-full">
+                {#if activeReq.self.status === "Idle"}
+                    <div class="bg-card flex size-full items-center justify-center gap-2 pb-8">
+                        <RocketIcon size="20" />
+                        <span>
+                            Hit <b class="font-semibold">Send</b> to make a request
+                        </span>
+                    </div>
+                {:else if activeReq.self.status === "Pending"}
+                    <div class="flex size-full items-center justify-center">
+                        <RefreshCwIcon
+                            strokeWidth={1.5}
+                            absoluteStrokeWidth
+                            size={20}
+                            class="mr-3 animate-spin"
+                        />
+                    </div>
+                {:else if activeReq.self.status === "Success" || activeReq.self.status === "Error"}
+                    <Tabs value="pretty" class="bg-card size-full">
+                        <div class="flex items-center justify-end border-b px-3">
+                            <TabsList class="my-1 auto-cols-min grid-flow-col gap-2 p-0">
+                                <TabsTrigger value="pretty">Pretty</TabsTrigger>
+                                <TabsTrigger value="raw">Raw</TabsTrigger>
+                                <TabsTrigger value="preview">Preview</TabsTrigger>
+                            </TabsList>
+                        </div>
+                        <div class="h-[calc(100%-2.25rem)] w-full overflow-scroll [&>*]:m-0">
+                            <TabsContent value="pretty" class="size-full">
+                                <CodeBlock
+                                    language={json()}
+                                    readOnly={true}
+                                    value={prettyJson(activeReq.self.response?.data)}
+                                    class="size-full"
+                                />
+                            </TabsContent>
+                            <TabsContent value="raw" class="size-full">
+                                <CodeBlock
+                                    language={null}
+                                    readOnly={true}
+                                    value={activeReq.self.response?.data}
+                                    class="size-full"
+                                />
+                            </TabsContent>
+                            <TabsContent value="preview" class="size-full">
+                                <iframe
+                                    title=""
+                                    src="about:blank"
+                                    srcdoc={activeReq.self.response
+                                        ? activeReq.self.response.data
+                                        : ""}
+                                    class="size-full"
+                                    loading="lazy"
+                                    sandbox=""
+                                ></iframe>
+                            </TabsContent>
+                        </div>
+                    </Tabs>
+                {/if}
+            </TabsContent>
+            <TabsContent value="cookies" class="m-0 size-full">
+                {#if activeReq.self.response}
+                    {@render cookiesTable(activeReq.self.response.cookies)}
+                {:else}
+                    <div class="flex size-full items-center justify-center">
+                        No cookies for you :(
+                    </div>
+                {/if}
+            </TabsContent>
+            <TabsContent value="headers" class="m-0 size-full">
+                {#if activeReq.self.response}
+                    {@render headersTable(activeReq.self.response.headers)}
+                {:else}
+                    <div class="flex size-full items-center justify-center">
+                        No headers received
+                    </div>
+                {/if}
+            </TabsContent>
+        </div>
     {/if}
-</div>
+</Tabs>
